Update: RealCrossing: Modularization and improvement of prediction (now on by default).
