+++ src/main/java/org/opennars/applications/crossing/Crossing.java
@@ -119,7 +119,9 @@ public class Crossing extends PApplet {
         }
 
         for (Entity e : entities) {
-            e.draw(this, streets, trafficLights, entities, null, 0);
+            e.draw(this, null, 0);
+            e.simulate(trafficLights, entities, streets);
+            
         }
         for (TrafficLight tl : trafficLights) {
             tl.draw(this, t);
@@ -137,7 +139,7 @@ public class Crossing extends PApplet {
         removeOutdatedPredictions(disappointments);
         for (Prediction pred : predictions) {
             Entity e = pred.ent;
-            e.draw(this, streets, trafficLights, entities, pred.truth, pred.time - nar.time());
+            e.draw(this, pred.truth, pred.time - nar.time());
         }
         if(showAnomalies) {
             for (Prediction pred : disappointments) {
@@ -149,7 +151,7 @@ public class Crossing extends PApplet {
                     fill(0,0,255);
                 }
                 this.text("ANOMALY", (float)e.posX, (float)e.posY);
-                e.draw(this, streets, trafficLights, entities, pred.truth, pred.time - nar.time());
+                e.draw(this, pred.truth, pred.time - nar.time());
             }
         }
         for(Camera c : cameras) {
