+++ src/main/java/org/opennars/applications/crossing/OperatorPanel.form
@@ -180,6 +180,7 @@
     </Container>
     <Component class="javax.swing.JCheckBox" name="jCheckBox1">
       <Properties>
+        <Property name="selected" type="boolean" value="true"/>
         <Property name="text" type="java.lang.String" value="Show predictions"/>
       </Properties>
       <Events>
